Fin Paciente y validaciones

+++ Presentacion/frmEspecialidad.cs
@@ -27,7 +27,15 @@ namespace Presentacion
 
         private void button4_Click(object sender, EventArgs e)
         {
-
+            errorProvider1.Clear();
+            if (tbBuscar.Text.Equals(""))
+            {
+                MessageBox.Show("Campo de busqueda vacio", "Advertencia", MessageBoxButtons.OK, MessageBoxIcon.Stop);
+            }
+            else
+            {
+                buscar();
+            }
         }
 
         private void groupBox1_Enter(object sender, EventArgs e)
@@ -69,13 +77,18 @@ namespace Presentacion
 
         private void button3_Click(object sender, EventArgs e)
         {
+            
+            errorProvider1.Clear();
             try
             {
-                Especialidad especialidad = new Especialidad();
-                especialidad.nombre = tbNombre.Text.Trim();
-                LN.agregarEspecialidad(especialidad);
-                limpiarDatos();
-                cargarDatos();
+                if (validarDatos(2))
+                {
+                    Especialidad especialidad = new Especialidad();
+                    especialidad.nombre = tbNombre.Text.Trim();
+                    LN.agregarEspecialidad(especialidad);
+                    limpiarDatos();
+                    cargarDatos();
+                }
             }
             catch (Exception)
             {
@@ -85,6 +98,7 @@ namespace Presentacion
         }
         private void limpiarDatos()
         {
+            errorProvider1.Clear();
             this.tbNombre.Text = string.Empty;
             this.tbIdEspecialidad.Text = string.Empty;
         }
@@ -93,12 +107,16 @@ namespace Presentacion
         {
             try
             {
-                Especialidad especialidad = new Especialidad();
-                especialidad.idEspecialidad = Convert.ToInt32(tbIdEspecialidad.Text.Trim());
-                especialidad.nombre = tbNombre.Text.Trim();
-                LN.modificarEspecialidad(especialidad);
-                limpiarDatos();
-                cargarDatos();
+                if (validarDatos(1))
+                {
+                    errorProvider1.Clear();
+                    Especialidad especialidad = new Especialidad();
+                    especialidad.idEspecialidad = Convert.ToInt32(tbIdEspecialidad.Text.Trim());
+                    especialidad.nombre = tbNombre.Text.Trim();
+                    LN.modificarEspecialidad(especialidad);
+                    limpiarDatos();
+                    cargarDatos();
+                }
             }
             catch (Exception)
             {
@@ -110,12 +128,21 @@ namespace Presentacion
         {
             try
             {
-                Especialidad especialidad = new Especialidad();
-                especialidad.idEspecialidad = Convert.ToInt32(tbIdEspecialidad.Text.Trim());
-                LN.eliminarEspecialidad(especialidad);
-                limpiarDatos();
-                cargarDatos();
-            }
+                errorProvider1.Clear();
+                if (tbIdEspecialidad.Text.Length <= 0)
+                {
+                    MessageBox.Show("Id Especialidad Vacio", "Advertencia", MessageBoxButtons.OK, MessageBoxIcon.Stop);
+                }
+                else
+
+                {
+                    Especialidad especialidad = new Especialidad();
+                    especialidad.idEspecialidad = Convert.ToInt32(tbIdEspecialidad.Text.Trim());
+                    LN.eliminarEspecialidad(especialidad);
+                    limpiarDatos();
+                    cargarDatos();
+                }
+                }
             catch (Exception)
             {
                 throw;
@@ -155,20 +182,14 @@ namespace Presentacion
 
         private void tbBuscar_TextChanged(object sender, EventArgs e)
         {
-            if (tbBuscar.Text.Equals(""))
-            {
-                MessageBox.Show("Campo de busqueda vacio", "Advertencia", MessageBoxButtons.OK, MessageBoxIcon.Stop);
-            }
-            else
-            {
-                buscar();
-            }
+           
         }
 
         private void dgridEspecialidad_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             try
             {
+                errorProvider1.Clear();
                 if (e.RowIndex < 0)
                 {
                     MessageBox.Show("Hizo click en una fila no permitida", "Advertencia", MessageBoxButtons.OK, MessageBoxIcon.Stop);
@@ -189,7 +210,43 @@ namespace Presentacion
 
         private void button5_Click(object sender, EventArgs e)
         {
+            errorProvider1.Clear();
             cargarDatos();
+        }
+
+        private void btLimpiar_Click(object sender, EventArgs e)
+        {
+            errorProvider1.Clear();
+            limpiar();
+        }
+        public void limpiar()
+        {
+            tbBuscar.Text = string.Empty;
+            tbIdEspecialidad.Text = string.Empty;
+            tbNombre.Text = string.Empty;
+        }
+        private bool validarDatos(int tipo)
+        {
+            int cont = 0;
+            errorProvider1.Clear();
+            if (tbIdEspecialidad.Text.Length <= 0 && tipo == 1)// si es uno valida Modificar, si es diferente a 1 valida Eliminar
+            {
+                errorProvider1.SetError(tbIdEspecialidad, "Id vacia");
+                cont++;
+            }
+            if (tbNombre.Text.Length <= 0)
+            {
+                errorProvider1.SetError(tbNombre, "Chofer vacio");
+                cont++;
+            }
+            if (cont == 0)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
     }
 }
